added --images-glob and added feedback on training images
Added --images-glob so that training images could be specified explicitly.
Still defaults to 'dataset/*/*.jpg' as before.

Added a check that the number of training files is not zero. Now issues
an error instead of going into infinite loop. The number of training images
found is also printed out to the console.